+++ src/routers/login/router.ts
@@ -2,7 +2,7 @@ import { Router } from 'express';
 
 export const loginRouter = Router();
 
-loginRouter.route('/login')
+loginRouter.route('/')
   //TODO: Replace by POST
   .get((req, res) => {
     res.send('Login success!');
